+++ core/services/offchainreporting/peer_wrapper.go
@@ -1,12 +1,14 @@
 package offchainreporting
 
 import (
+	"strings"
 	"sync"
 
 	"github.com/jinzhu/gorm"
 	"github.com/pkg/errors"
 	"github.com/smartcontractkit/chainlink/core/logger"
 	"github.com/smartcontractkit/chainlink/core/store/models"
+	"github.com/smartcontractkit/chainlink/core/store/models/p2pkey"
 	"github.com/smartcontractkit/chainlink/core/store/orm"
 	ocrnetworking "github.com/smartcontractkit/libocr/networking"
 	ocrtypes "github.com/smartcontractkit/libocr/offchainreporting/types"
@@ -65,11 +67,35 @@ func (p *SingletonPeerWrapper) Start() (err error) {
 	if len(p2pkeys) == 0 {
 		return nil
 	}
-	if len(p2pkeys) > 1 {
-		return errors.New("more than one p2p key is not currently supported")
+
+	var key p2pkey.Key
+	var matched bool
+	checkedKeys := []string{}
+	configuredPeerID, err := p.config.P2PPeerID(nil)
+	if err != nil {
+		return errors.Wrap(err, "failed to start peer wrapper")
+	}
+	for _, k := range p2pkeys {
+		var peerID models.PeerID
+		peerID, err = k.GetPeerID()
+		if err != nil {
+			return errors.Wrap(err, "unexpectedly failed to get peer ID from key")
+		}
+		if peerID == configuredPeerID {
+			key = k
+			matched = true
+			break
+		}
+		checkedKeys = append(checkedKeys, peerID.String())
+	}
+	keys := strings.Join(checkedKeys, ", ")
+	if !matched {
+		if configuredPeerID == "" {
+			return errors.Errorf("multiple p2p keys found but peer ID was not set. You must specify P2P_PEER_ID if you have more than one key. Keys available: %s", keys)
+		}
+		return errors.Errorf("multiple p2p keys found but none matched the given P2P_PEER_ID of '%s'. Keys available: %s", configuredPeerID, keys)
 	}
 
-	key := p2pkeys[0]
 	p.PeerID, err = key.GetPeerID()
 	if err != nil {
 		return errors.Wrap(err, "could not get peer ID")